+++ docs_jekyll/build.sh
@@ -4,7 +4,15 @@ echo "Copying screenshots ... "
 cp -R ../screenshots/ .
 
 echo "Building markdown page from template and readme ..."
-cat __partials/index-head.partial.html ../README.md __partials/index-foot.partial.html > index.md
+echo '---' > index.md
+echo 'layout: home' >> index.md
+echo 'date: 2019-03-11' >> index.md
+echo '---' >> index.md
+cat ../README.md >> index.md
+echo '## Changelog' >> index.md
+echo '[View changelog](changelog.html "View changelog")' >> index.md
+
+# cat __partials/index-head.partial.html ../README.md __partials/index-foot.partial.html > index.md
 
 echo "Building the changelog file ..."
 echo '---' > changelog.md