+++ src/index.js
@@ -59,7 +59,7 @@
 	                         }, 
 				 log : function() {
 				     console.log(arguments);
-				     humane.log(arguments[0]);
+				     // humane.log(arguments[0]);
 				 },
 				 error : function() {
 				     console.error(arguments);
@@ -67,7 +67,7 @@
 				 }
 			       } );
 	    }
-	    humane.log('PlotBoilerplate');
+	    humane.log('plotboilerplate');
 	    
 	    // +---------------------------------------------------------------------------------
 	    // | Initialize dat.gui